% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{chunk_users}
\alias{chunk_users}
\alias{chunk_users_data}
\title{Chunk users}
\usage{
chunk_users(x, n = 50)

chunk_users_data(x, n = 50)
}
\arguments{
\item{x}{Input vector of users. Duplicates and missing values will be removed}

\item{n}{Number of users per chunk (users returned in each element of output)}
}
\value{
chunk_users: returns a list containing character vectors

chunk_users_data: returns a list containing data frames
}
\description{
Convert an atomic vector of users into a list of atomic vectors
}
\examples{
## this generates a vector of user-ID like values
users <- replicate(1000, paste(sample(0:9, 14, replace = TRUE), collapse = ""))

## break users into 100-user chunks
chunky <- chunk_users(users, n = 100)

## preview returned object
str(chunky, 1)

}
