% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{log_counts}
\alias{log_counts}
\title{Log counts}
\usage{
log_counts(x)
}
\arguments{
\item{x}{Input data}
}
\value{
Output should match input class
}
\description{
Safely (deals with zero and negative values) logs integers
}
\keyword{internal}
