% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict_bot}
\alias{predict_bot}
\title{Predict Twitter bots}
\usage{
predict_bot(x)
}
\arguments{
\item{x}{Input data either character vector of Twitter identifiers (user IDs
or screen names) or a data frame of Twitter data}
}
\value{
A data frame with the user id, screen name, and estimated probability
of being a bot
}
\description{
Estimate probability that one or more Twitter accounts is a "bot"
}
