% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{preprocess_bot}
\alias{preprocess_bot}
\title{Preprocess data}
\usage{
preprocess_bot(x, batch_size = 100)
}
\arguments{
\item{x}{Input either a user identifier or Twitter data}
}
\value{
Returns a data frame used to generate predictions
}
\description{
Prepares data for Twitter bot model
}
