% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{sample_via_twitter_lists}
\alias{sample_via_twitter_lists}
\title{Sample Twitter Accounts}
\usage{
sample_via_twitter_lists(x, n = 1000, regex, ...)
}
\arguments{
\item{x}{Vector or data frame containing screen names/user IDs. Or a vector
or data frame containing list IDs/uri's.}

\item{n}{Maximum number of sample users to return.}

\item{regex}{Regular expression string used to filter lists–this pattern
will be used to select which lists are relevant, e.g., "bot" will only use
lists that contain the consecutive letters "bot"}

\item{...}{Other args passed to rtweet function calls}
}
\value{
Returns a data table with user_id, screen_name, description, and n
(list appearance count)
#' @details
The search-based sampling method finds accounts by querying Twitter's
'users/search' endpoint, which provides a simple interface for searching
public user profile information.

Build a dataset by first selecting accounts using
\enumerate{
\item user via lists_users & lists_members
\item list via lists_members
}
}
\description{
Uses a snowball sampling method to identify (and gather) Twitter accounts
}
