% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{split_test_train}
\alias{split_test_train}
\title{Split test train}
\usage{
split_test_train(.data, .p = 0.8, ...)
}
\arguments{
\item{.data}{Input data set}

\item{.p}{Proportion of cases to appear in training data}

\item{...}{Optional, specify response variable via non-standard evaluation}
}
\value{
a list with "train" and "test" data frames
}
\description{
Splits data frame into train and test sets
}
\keyword{internal}
