% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweetbotornot2-package.R
\docType{package}
\name{tweetbotornot2-package}
\alias{tweetbotornot2}
\alias{tweetbotornot2-package}
\title{tweetbotornot2: Detect Twitter Bots}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Providing an out-of-the-box classifier for detecting Twitter bots that
    is easy to use, interpretable, scalable, and accurate.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mkearney/tweetbotornot2}
  \item Report bugs at \url{https://github.com/mkearney/tweetbotornot2/issues}
}

}
\author{
\strong{Maintainer}: Michael W. Kearney \email{kearneymw@missouri.edu} (\href{https://orcid.org/0000-0002-0730-4694}{ORCID})

}
\keyword{internal}
