% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explain.R
\name{tweetbotornot_explain}
\alias{tweetbotornot_explain}
\title{Explain Twitter bot predictions}
\usage{
tweetbotornot_explain(x)
}
\arguments{
\item{x}{Input data either character vector of Twitter identifiers (user IDs
or screen names), data frame of Twitter data, or predictions returned by
tweetbotornot_explain}
}
\value{
A data frame with the user id, screen name, probability estimate,
feature name, and feature contribution
}
\description{
Explain estimated probability that one or more Twitter accounts is a "bot"
}
\examples{

\dontrun{

## estimate prediction and return with feature contribution
kmw <- tweetbotornot_explain("kearneymw")

## view data
kmw

## prob_bot should be roughly equal to sum of log-odds of 'value'
kmw[,
  .(prob_bot = stats::qlogis(prob_bot[1]),
  contr = sum(value)),
  by = screen_name
 ]
}
}
