% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/botometer.R
\name{predict_botometer}
\alias{predict_botometer}
\title{Botometer scores}
\usage{
predict_botometer(users, token = NULL, key = NULL, set_key = FALSE)
}
\arguments{
\item{users}{Vector of user IDs or screen names}

\item{token}{rtweet token. If NULL (default), then rtweet looks for a token
path in .Renviron}

\item{key}{Botometer API key, which you should be able to get from:
https://rapidapi.com/OSoMe/api/botometer/details}

\item{set_key}{Logical indicating whether to set the botometer key as an
R environment variable for current future sessions (on the same machine).}
}
\value{
A data frame (data.table) of user ID, screen name, and botometer score
}
\description{
Get Botometer scores for given user(s)
}
