% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/botspotr-package.R
\docType{package}
\name{tweetbotornot2-package}
\alias{tweetbotornot2}
\alias{tweetbotornot2-package}
\title{tweetbotornot2: Detect Twitter Bots}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

This package provides an out-of-the-box classifier for detecting Twitter bots
    that is easy to use, volume-friendly (classify over two hundred thousands accounts per day),
    accurate (see paper), robust and customizable (update the model with new or user-defined training data),
    and explainable (view contributions for features used in estimates).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mkearney/tweetbotornot2}
  \item Report bugs at \url{https://github.com/mkearney/tweetbotornot2/issues}
}

}
\author{
\strong{Maintainer}: Michael W. Kearney \email{kearneymw@missouri.edu} (\href{https://orcid.org/0000-0002-0730-4694}{ORCID})

}
\keyword{internal}
